kiro/voiquyr/voiquyr-command-center/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
deepgram-sdk==3.1.0
websockets==12.0
redis==5.0.1
openai==1.12.0
elevenlabs==0.2.27
requests==2.31.0
aiohttp==3.9.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9